% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackagef20-package.R
\docType{package}
\name{testpackagef20-package}
\alias{testpackagef20}
\alias{testpackagef20-package}
\title{testpackagef20: My Package as a Demo Package)}
\description{
What the package does still tbd.
}
\author{
\strong{Maintainer}: Richard Ressler \email{rressler@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
